requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
langchain==0.3.7
langchain_community==0.3.7
python-dotenv==1.0.1
langchain_google_genai==2.0.5
huggingface_hub-0.26.2